Change Date into Settlement Date, and add proper USD/CAD lookup for bank holidays
Each TX now has a separate TradeDate and SettlementDate, since they
are independently useful. The settlement date is used for tax
year determination, while the trade date is mainly just used (for our
purposes anyway) to get the foreign exchange rate used at the time
of the trade.

I took this opportunity to properly look up what rate is supposed to be
used, and came up with a reasonably confident interpretation, which
seems to make logical sense.