Allow specified SFLs to be forced
In cases where the transaction(s) inducing a superficial loss is not
tracked, the user may want to force a specific superficial loss
adjustment, assuming they know what they are doing.

This can now be done by adding an ! to the end of the specified loss
value. Without it, acb will assume the value should be exactly the
calculated SFL and emit an error if it is not. This is since the manual
SFL is used to indicate that the ACB adjust will also be manually entered,
and any error in the SFL would likely mean the ACB adjustment entries are
also incorrect.