acb: add tracking of total costs across all securities (#25)
* app: add CSV output option by implementing the ACBWriter interface

This will write a single CSV file with the ACB for each security.
Basically, any time a new RenderTable is printed, it will create a
separate CSV file.

* acb: add tracking of total costs across all securities

This keeps track of the total cost basis across all securities for
every day where any trades have been made. It will also use this
information to present a yearly summary of the maximum cost of all
tracked securities in aggregate. This is useful in determining if, for
example, a T1135 form will be required for a given year.

Since it now generates multiple reports, there's a new
--csv-output-dir that will cause all of the reports to be written to
separate CSV files in the specified directory for easier consumption.

A new test, test/total_cost_test.go, checks the correctness
of the aggregate cost reporting.

* acb: handle txDelta from registered affiliate in RenderTotalCosts

These transactions have a null value for TotalAcb, and so including
them would invalidate our daily, and then yearly, totals. These should
simply be excluded.

Notes will be added to the rendered table for each instance of the
same.

Note that if there are only registered transactions for a given
security that it will not show up in the costs table at all (outside
of the warnings / notes at the bottom).

* acb: handle non-default affiliate in RenderTotalCosts

When calculating the total costs, we need to also skip non-default
affiliates for unregistered affiliate transactions. This needs an
associated warning as well. Added this to the TestRenderTotalCosts
test cases.

I've adjusted the other integration tests to re-run their test cases
with and without "renderTotalCosts" as an additional smoke test.

---------

Co-authored-by: wwade <wwade@users.noreply.github.com>