etrade: fix handling of ESO confirmation with multiple grants
The assumed format was incorrect. It actually presents the multiple
grants as separate columns with only a row prefix, that is, e.g.:

    Grant 1 Grant 2
    Grant Number 1234 1235

This is the reason `search_for_rows` had a `rowvalue2`. In theory,
this could probably be extended to "n" Grant columns, but we don't
have any test data for that case.